refactor : baseEntity, loginUser
