--- src/main/java/kr/co/pinup/common/BaseEntity.java
@@ -1,27 +1,0 @@
-package kr.co.pinup.common;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Long id;
-
-    @CreatedDate
-    @Column(updatable = false)
-    protected LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(insertable = false)
-    protected LocalDateTime updatedAt;
-}
