+++ src/main/java/kr/co/pinup/faqs/Faq.java
@@ -1,7 +1,7 @@
 package kr.co.pinup.faqs;
 
 import jakarta.persistence.*;
-import kr.co.pinup.common.BaseEntity;
+import kr.co.pinup.BaseEntity;
 import kr.co.pinup.faqs.model.enums.FaqCategory;
 import kr.co.pinup.faqs.model.dto.FaqUpdateRequest;
 import kr.co.pinup.members.Member;
