+++ src/main/java/kr/co/pinup/notices/Notice.java
@@ -1,7 +1,7 @@
 package kr.co.pinup.notices;
 
 import jakarta.persistence.*;
-import kr.co.pinup.common.BaseEntity;
+import kr.co.pinup.BaseEntity;
 import kr.co.pinup.notices.model.dto.NoticeUpdateRequest;
 import kr.co.pinup.members.Member;
 import lombok.*;
