+++ src/main/resources/templates/index.html
@@ -27,4 +27,4 @@
 </th:block>
 </html>
 
-<script src="/js/Member.js"></script>
+<script src="/js/member.js"></script>
