--- src/test/resources/application-test.yml
@@ -1,109 +1,0 @@
-spring:
-  application:
-    name: pinup-test
-  main:
-    allow-bean-definition-overriding: true
-
-  datasource:
-    url: jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
-    driverClassName: org.h2.Driver
-    username: sa
-    password:
-    hikari:
-      maximum-pool-size: 10
-
-      h2:
-        console:
-          enabled: true
-          path: /h2-console
-
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    show-sql: true
-    database-platform: org.hibernate.dialect.H2Dialect
-    properties:
-      hibernate:
-        format_sql: true
-
-  thymeleaf:
-    prefix: classpath:/templates/
-    suffix: .html
-    mode: HTML
-    encoding: UTF-8
-    cache: false
-
-  security:
-    oauth2:
-      client:
-        registration:
-          naver:
-            client-id: test-naver-id
-            client-secret: test-naver-secret
-            scope:
-              - name
-              - email
-            client-name: Naver
-            authorization-grant-type: authorization_code
-            redirect-uri: http://localhost:8888/test/members/oauth/naver
-          google:
-            client-id: test-google-id
-            client-secret: test-google-secret
-            scope:
-              - profile
-              - email
-            client-name: Google
-            authorization-grant-type: authorization_code
-            redirect-uri: http://localhost:8888/test/members/oauth/google
-
-        provider:
-          naver:
-            authorization-uri: http://localhost:8888/test/members/naver/authorize
-            token-uri: http://localhost:8888/test/members/naver/token
-            user-info-uri: http://localhost:8888/test/members/naver/userInfo
-            user-name-attribute: response
-          google:
-            authorization-uri: http://localhost:8888/test/members/google/authorize
-            token-uri: http://localhost:8888/test/members/google/token
-            user-info-uri: http://localhost:8888/test/members/google/userInfo
-            user-name-attribute: sub
-
-springdoc:
-  packages-to-scan: com.colabear754.springdoc_example.controllers
-  default-consumes-media-type: application/json;charset=UTF-8
-  default-produces-media-type: application/json;charset=UTF-8
-  swagger-ui:
-    path: /springdoc.html
-    disable-swagger-default-url: true
-    display-request-duration: true
-    operations-sorter: alpha
-
-server:
-  servlet:
-    context-path: /
-    encoding:
-      enabled: true
-      charset: UTF-8
-      force: true
-  port: 8888
-
-logging:
-  level:
-    root: INFO  # 전체 애플리케이션의 기본 로그 레벨을 INFO로 설정
-    org.springframework.web: DEBUG  # Spring Web 관련 로그는 DEBUG로 설정
-    org.springframework.security: DEBUG  # Spring Security 관련 로그는 DEBUG로 설정
-    org.hibernate.SQL: DEBUG
-    org.hibernate.type.descriptor.sql.BasicBinder: TRACE
-    com.colabear754: TRACE  # 사용자 패키지(com.colabear754) 관련 로그는 TRACE로 설정 (세밀한 로그 기록)
-    org.springframework.boot: INFO  # Spring Boot 관련 로그는 INFO 레벨로 설정
-    org.springframework.context: DEBUG  # Spring Context 관련 로그는 DEBUG로 설정
-#    com.zaxxer.hikari: DEBUG  # HikariCP 관련 로그는 DEBUG로 설정 (데이터베이스 커넥션 풀 정보)
-
-cloud:
-  aws:
-    s3:
-      endpoint: http://192.168.45.76:4566  # LocalStack S3 엔드포인트
-      region: us-east-1
-    credentials:
-      access-key: test
-      secret-key: test
